Fix memory leak in CustomPayloadS2CPacket
This is technically a bug in vanilla, however vanilla is saved from this bug due to pure dumb luck: it extensively abuses unpooled heap buffers, so the buffer just gets picked up by the garbage collector and we have nothing to worry about. However, with pooled direct buffers (as with Krypton), the copy is a big problem since it will occupy precious direct buffer space. Worse still, the copy isn't even needed at all. Change it to a slice, which is what it should've been in the first place.

While we're at it, add a new mixin to allow the resource leak detection level to be changed via the normal system property.