Add system property to permit oversized packets.
This should fix #91, but this is not enabled by default for security reasons. This also does not raise all relevant packet lengths capped by the protocol, although this can be added at a future point in time.

With this system property enabled, Krypton permits packets with 5-byte VarInt prefixes and allows uncompressed packets to be up to 2GiB in size.